use generic test case for login/register
