+++ server/tests/test_routes_api.py
@@ -16,103 +16,31 @@ from jsonschema import validate, ValidationError
 from flask import session
 from flask_security import current_user, url_for_security
 
-from ludobox import create_app
 from ludobox.run import get_server_port
-from ludobox.config import read_config
 from ludobox.content import read_content
 from ludobox.models import User, Role, db
 from ludobox.security import user_datastore
 
 from ludobox.routes.api import rest_api
 
+# test helpers
+from LudoboxTestCase import LudoboxTestCase
 from helpers import delete_data_path, create_empty_data_path, add_samples_to_data_dir
 
 TEST_DATA_DIR = '/tmp/test-data'
 
-class TestLudoboxWebServer(TestCase):
-
-
-    def create_app(self):
-        # pass in test configuration
-        test_dir = os.path.join(os.getcwd(),"server/tests")
-        config_path = os.path.join(test_dir,"config.test.yml")
-
-        # create empy path for data
-        delete_data_path(TEST_DATA_DIR)
-        create_empty_data_path(TEST_DATA_DIR)
-
-        return create_app(self, config_path=config_path)
+class TestLudoboxWebServer(LudoboxTestCase):
 
     def setUp(self):
-
-        self.config = read_config()
-        self.app.config['SECRET_KEY'] = 'sekrit!'
-        self.app.config["WTF_CSRF_ENABLED"] = False
-        self.app.config["LOGIN_DISABLED"] = False
-
-        # create games
-        add_samples_to_data_dir(self.app.config["DATA_DIR"])
 
         # register routes
         self.app.register_blueprint(rest_api)
 
-        # setup db
-        db.drop_all()
-        db.create_all()
-
-        # create default roles
-        user_role = Role(name='contributor')
-        super_user_role = Role(name='superuser')
-        db.session.add(user_role)
-        db.session.add(super_user_role)
-        db.session.commit()
-
-        # add a role
-        # default_role = user_datastore.find_role("contributor")
-        # user_datastore.add_role_to_user(self.user, default_role)
-        # db.session.commit()
-
-        # add a user
-        self.user_email = "tester@test.com"
-        self.user_password = "password"
+        # register a new user
         rv = self.register(
-            email=self.user_email, password=self.user_password
+            email=self.user_email,
+            password=self.user_password
             )
-
-        # creates a test client
-        self.client = self.app.test_client()
-
-        # propagate the exceptions to the test client
-        self.client.testing = True
-
-    def tearDown(self):
-        db.session.remove()
-
-    def register(self, email=None, password=None):
-        """Register a user"""
-        print email, password
-        return self.client.post(
-            url_for_security('register'),
-            data={
-                'email': email,
-                'password': password,
-                'password_confirm': password
-                },
-            content_type= 'application/x-www-form-urlencoded',
-            follow_redirects=True
-        )
-
-    def login(self, email=None, password=None):
-        email = email or self.user_email
-        password = password or self.user_password
-        return self.client.post(
-            url_for_security('login'),
-            data={'email': email, 'password': password},
-            follow_redirects=True
-        )
-
-    def logout(self):
-        return self.client.get('/logout', follow_redirects=True)
 
     def test_login_logout(self):
         rv = self.login(email=None, password=None)
@@ -139,28 +67,6 @@ class TestLudoboxWebServer(TestCase):
 
         self.assertEqual(result.status_code, 200)
         # self.assertEqual(json.loads(result.data), {"name" : self.config["ludobox_name"]})
-
-    def test_css(self):
-        result = self.client.get('/css/style.css')
-        self.assertEqual(result.status_code, 200)
-
-    def test_js(self):
-        result = self.client.get('/js/bundle.js')
-        self.assertEqual(result.status_code, 200)
-
-    def test_images(self):
-        result = self.client.get('/images/favicon.png')
-        self.assertEqual(result.status_code, 200)
-
-    def test_game_schema(self):
-        result = self.client.get('/api/schema/game')
-        self.assertEqual(result.status_code, 200)
-
-        schema = json.loads(result.data)
-        self.assertIs(type(schema), dict)
-        print schema
-
-        self.assertRaises(ValidationError, lambda : validate({ "test" : [1,3,2,4]}, schema))
 
     def test_files_api_with_no_files(self):
         result = self.client.get('/api/files/dsdqdqs')
@@ -192,7 +98,12 @@ class TestLudoboxWebServer(TestCase):
             self.assertEqual(result.status_code, 401)
 
     def test_add_game_forbidden(self):
-        valid_info = read_content(os.path.join(os.getcwd(), 'server/tests/test-data/test-game'))
+        """Make sure only logged in user can add content"""
+
+        valid_info = self.borgia_info_content
+
+        delete_data_path(self.app.config["DATA_DIR"])
+        create_empty_data_path(self.app.config["DATA_DIR"])
 
         data = {
             'files': [
@@ -203,12 +114,17 @@ class TestLudoboxWebServer(TestCase):
             'info': json.dumps(valid_info)
         }
 
-        result = self.client.post('/api/create',
-                                data=data,
-                                content_type='multipart/form-data'
-                                )
+        with self.app.test_client() as c:
 
-        self.assertEqual(result.status_code, 403)
+            print current_user
+
+            result = c.post('/api/create',
+                                    data=data,
+                                    content_type='multipart/form-data'
+                                    )
+
+            print result.data
+            self.assertEqual(result.status_code, 403)
 
     def test_api_create_content(self):
